Work around race condition bug in tmux
It looks like tmux doesn't wait for me to finish defining my
default-terminal before it opens up my new shell. This means the first
window of my new tmux sessions usually humbly thinks it only has eight
colours.

This seems to be a known bug in tmux, so for the moment I'm just working
around it with an environment variable with which .bashrc will deal.