Start using URxvt clipboard script
This script from urxvt-perls[1] copies the PRIMARY X11 selection to the
CLIPBOARD. I do like select-to-copy and middle-click pasting most of the
time, but it's occasionally frustrating when doing things like trying to
replace selected text. This allows me to do explicit pastes with Ctrl-V.

I don't like having to include my $HOME path literal in .Xresources;
I'll need to find out if there's a tidier way to do it.

[1]: https://github.com/muennich/urxvt-perls